# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Amazon Prime Video API – Spécification OpenAPI 3.0
  description: |
    Spécification complète simulant l'API d'Amazon Prime Video.

    ## Domaines couverts
    - **Auth** — Connexion, tokens JWT, 2FA
    - **Profils** — Gestion multi-profils (comme Netflix)
    - **Catalogue** — Films, séries, épisodes, documentaires
    - **Recherche** — Recherche textuelle, par genre, par acteur
    - **🔍 Recherche par image (IA)** — Identifier un film/acteur depuis une photo
    - **Lecteur** — Lecture, reprise, qualité vidéo, sous-titres
    - **Watchlist** — Liste à regarder plus tard
    - **Historique** — Contenus vus et reprises automatiques
    - **Évaluations & Avis** — Notes et commentaires
    - **Téléchargements** — Visionnage hors ligne
    - **Abonnement** — Prime Video, chaînes, add-ons
    - **Paiement** — Transactions, factures
    - **Recommandations** — Personnalisées, tendances, similaires
    - **Notifications** — Nouvelles sorties, rappels
    - **Contrôle parental** — Restrictions par profil
    - **📺 Publicités IA** — Génération automatique de bandes-annonces et spots publicitaires par IA
    - **Live & Events** — Streaming en direct, sport, événements
    - **Langues & Accessibilité** — Sous-titres, audiodescription, multi-langue

    ## Authentification
    Utiliser `POST /auth/login` pour obtenir un `accessToken` JWT (durée 1h).
    À inclure dans toutes les requêtes protégées : `Authorization: Bearer {accessToken}`

...
  contact:
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenu via `POST /auth/login`. À inclure dans le header : `Authorization: Bearer {accessToken}`. Durée de validité : 3600 secondes (1 heure).
  schemas:
    Error:
      type: object
      description: |
        Réponse d'erreur standard. `code` est machine-readable (constante), `message` est lisible par un humain, `details` liste les erreurs de validation champ par champ.
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: RESOURCE_NOT_FOUND
        message:
          type: string
          example: Le contenu demandé est introuvable.
        details:
          type: array
          items:
            type: string
...
    LoginRequest:
      type: object
      description: |
        Identifiants de connexion. Transmis en HTTPS uniquement — jamais en HTTP non chiffré.
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: jean.dupont@email.com
        password:
          type: string
          format: password
          minLength: 8
          example: MonMotDePasse123!
        deviceId:
          type: string
          description: Identifiant unique de l'appareil (utilisé pour la gestion des appareils connectés)
          example: device_abc123
    LoginResponse:
      type: object
      description: |
        Tokens retournés après une connexion réussie. Utiliser `accessToken` dans le header `Authorization: Bearer {token}`. Le `refreshToken` permet de renouveler l'`accessToken` sans re-saisir le mot de passe.
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          example: dGhpcyBpcyBhIHZhbGlkIHJlZnJlc2ggdG9rZW4...
        expiresIn:
          type: integer
          description: Durée de validité de l'accessToken en secondes
          example: 3600
        tokenType:
          type: string
          example: Bearer
        requiresMFA:
          type: boolean
          description: Si true, un mfaToken est retourné à la place des tokens définitifs
        mfaToken:
          type: string
          nullable: true
          description: Token temporaire (5 min) à utiliser avec POST /auth/mfa/challenge si requiresMFA = true
    RegisterRequest:
      type: object
      description: Données nécessaires pour créer un nouveau compte Prime Video.
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        firstName:
          type: string
          example: Jean
        lastName:
          type: string
          example: Dupont
        phone:
          type: string
          example: '+33612345678'
        birthDate:
          type: string
          format: date
          example: '1990-05-15'
        country:
          type: string
          example: FR
    RefreshTokenRequest:
      type: object
      description: Payload pour renouveler un token d'accès expiré.
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    ForgotPasswordRequest:
      type: object
      description: Email pour déclencher une réinitialisation de mot de passe.
      required:
        - email
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      description: Token reçu par email + nouveau mot de passe.
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
        newPassword:
          type: string
          format: password
          minLength: 8
...
    MFASetup:
      type: object
      description: |
        Données de configuration 2FA : QR code à scanner avec Google Authenticator ou Authy, et codes de secours à conserver hors ligne en cas de perte du téléphone.
      properties:
        secret:
          type: string
          description: Clé secrète TOTP en Base32 (pour import manuel si QR indisponible)
          example: JBSWY3DPEHPK3PXP
        qrCodeUrl:
          type: string
          format: uri
          description: URL du QR code à scanner
        issuer:
          type: string
          example: Prime Video
        accountName:
          type: string
          example: jean.dupont@email.com
        backupCodes:
          type: array
          description: Codes de secours à usage unique — à sauvegarder hors ligne
          items:
            type: string
          example:
            - ABCD-1234
            - EFGH-5678
            - IJKL-9012
...
  responses:
    Unauthorized:
      description: Non authentifié — token manquant, invalide ou expiré
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            message: Token manquant ou invalide. Utilisez POST /auth/login pour obtenir un token.
...
    ValidationError:
      description: Erreur de validation — les données envoyées sont invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    TooManyRequests:
      description: Limite de taux dépassée — réessayer après le délai indiqué dans le header Retry-After
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Erreur interne du serveur
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
  /auth/register:
    post:
      tags:
        - Auth
      summary: Créer un compte Prime Video
      operationId: registerUser
      description: |
        Crée un nouveau compte utilisateur Prime Video.
        **Fonctionnement :** 1. Vérifie que l'email n'est pas déjà utilisé (sinon `409 Conflict`). 2. Hache le mot de passe avec bcrypt — jamais stocké en clair. 3. Envoie un email de vérification. 4. Retourne immédiatement des tokens JWT.
        **Règles mot de passe :** minimum 8 caractères, au moins une majuscule et un chiffre.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: jean.dupont@email.com
              password: MonMotDePasse123!
              firstName: Jean
              lastName: Dupont
              country: FR
      responses:
        '201':
          description: Compte créé — tokens retournés pour connexion immédiate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email déjà utilisé sur un compte existant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: EMAIL_ALREADY_EXISTS
                message: Un compte existe déjà avec cet email.
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/login:
    post:
      tags:
        - Auth
      summary: Connexion
      operationId: loginUser
      description: |
        Authentifie un utilisateur et retourne des tokens JWT.
        **Si le 2FA est activé :** la réponse contient `requiresMFA: true` et un `mfaToken` temporaire (5 min) au lieu des tokens définitifs. Appeler ensuite `POST /auth/mfa/challenge` pour compléter la connexion.
        **Sécurité :** limité à 10 tentatives par minute par IP (protection brute-force). Après 5 échecs consécutifs, un CAPTCHA est requis.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: jean.dupont@email.com
              password: MonMotDePasse123!
              deviceId: device_abc123
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Identifiants incorrects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: INVALID_CREDENTIALS
                message: Email ou mot de passe incorrect.
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Rafraîchir le token d'accès
      operationId: refreshToken
      description: |
        Génère un nouvel `accessToken` à partir d'un `refreshToken` valide.
        **Pourquoi ?** L'`accessToken` expire toutes les heures pour des raisons de sécurité. Au lieu de redemander le mot de passe, le client utilise silencieusement le `refreshToken` (valable 30 jours) pour obtenir un nouveau token.
        **Rotation automatique :** chaque appel invalide l'ancien `refreshToken` et en retourne un nouveau (refresh token rotation).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Nouveau token d'accès généré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Refresh token invalide ou expiré — reconnexion requise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/forgot-password:
    post:
      tags:
        - Auth
      summary: Demander la réinitialisation du mot de passe
      operationId: forgotPassword
      description: |
        Déclenche l'envoi d'un email de réinitialisation de mot de passe.
        **Sécurité :** la réponse est identique que l'email existe ou non dans la base de données — cela évite de révéler quels emails sont enregistrés sur la plateforme (attaque par énumération de comptes).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Email envoyé si le compte existe
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Si un compte correspond à cet email, un lien de réinitialisation vous a été envoyé.
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/reset-password:
    post:
      tags:
        - Auth
      summary: Réinitialiser le mot de passe
      operationId: resetPassword
      description: |
        Réinitialise le mot de passe via le token reçu par email.
        Le token est à usage unique et expire après 15 minutes. Une fois utilisé, il est immédiatement invalidé. Après modification, tous les `refreshToken` existants sont invalidés (déconnexion de tous les appareils).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Mot de passe réinitialisé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Token invalide ou expiré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: INVALID_TOKEN
                message: Le lien de réinitialisation est invalide ou a expiré.
...
  /auth/mfa/setup:
    post:
      tags:
        - Auth
      summary: Initialiser la configuration 2FA
      operationId: setupMFA
      description: |
        Initialise la configuration de l'authentification à deux facteurs (2FA/MFA).
        **Comment ça marche ?** 1. Le serveur génère une clé secrète TOTP unique pour cet utilisateur. 2. Un QR code est retourné — à scanner avec Google Authenticator ou Authy. 3. L'application génère un code à 6 chiffres renouvelé toutes les 30 secondes. 4. Le 2FA n'est pas encore actif — appeler `POST /auth/mfa/verify` avec
           le premier code pour confirmer et activer.

        Des **codes de secours** sont retournés — à conserver hors ligne en cas de perte du téléphone.
      responses:
        '200':
          description: Configuration 2FA initialisée — QR code et codes de secours retournés
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MFASetup'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Le 2FA est déjà activé sur ce compte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/mfa/verify:
    post:
      tags:
        - Auth
      summary: Activer le 2FA
      operationId: verifyMFA
      description: |
        Active définitivement le 2FA en validant le premier code TOTP généré par l'application.
        Après activation, chaque connexion nécessitera ce second facteur en plus du mot de passe.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
                  pattern: ^\d{6}$
                  example: '123456'
      responses:
        '200':
          description: 2FA activé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Code TOTP invalide ou expiré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/mfa/challenge:
    post:
      tags:
        - Auth
      summary: Valider le second facteur lors d'une connexion
      operationId: challengeMFA
      description: |
        Valide le second facteur lors d'une connexion sur un compte avec 2FA activé.
        **Flux complet de connexion avec 2FA :** 1. `POST /auth/login` → retourne `requiresMFA: true` et un `mfaToken` temporaire (5 min) 2. `POST /auth/mfa/challenge` avec `mfaToken` + code TOTP → retourne les tokens définitifs
        Accepte un code TOTP (6 chiffres, valide 30 sec) ou un code de secours (à usage unique).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - mfaToken
                - code
              properties:
                mfaToken:
                  type: string
                  description: Token temporaire reçu après POST /auth/login
                code:
                  type: string
                  description: Code TOTP (6 chiffres) ou code de secours
                  example: '123456'
      responses:
        '200':
          description: Authentification complète — tokens d'accès retournés
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Code invalide ou expiré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: mfaToken invalide ou expiré — relancer POST /auth/login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...